+++ packages/eslint-config-smarthr/index.js
@@ -55,8 +55,10 @@ export default [
       'smarthr/a11y-trigger-has-button': 'error',
       'smarthr/best-practice-for-async-current-target': 'error',
       'smarthr/best-practice-for-button-element': 'error',
+      'smarthr/best-practice-for-consecutive-definition-list': 'warn', // TODO: 2026/05にwarn化。問題なければerrorに変更予定
       'smarthr/best-practice-for-date': 'error',
       'smarthr/best-practice-for-data-test-attribute': 'off',
+      'smarthr/best-practice-for-default-props': 'warn', // TODO: 2026/05にwarn化。問題なければerrorに変更予定
       'smarthr/best-practice-for-interactive-element': 'error',
       'smarthr/best-practice-for-layouts': 'error',
       'smarthr/best-practice-for-nested-attributes-array-index': 'error',