+++ packages/eslint-config-smarthr/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [14.2.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.1.0...eslint-config-smarthr-v14.2.0) (2026-05-17)
+
+
+### Features
+
+* **eslint-config-smarthr:** eslint-plugin-smarthrを6.15.0に更新 ([#1316](https://github.com/kufu/tamatebako/issues/1316)) ([a549db3](https://github.com/kufu/tamatebako/commit/a549db3c09a0b1d270b4bec30f39c43b1e0c9891))
+
+## [14.1.0](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.0.1...eslint-config-smarthr-v14.1.0) (2026-05-14)
+
+
+### Features
+
+* **eslint-config-smarthr:** eslint-plugin-smarthrを6.14.0に更新 ([#1306](https://github.com/kufu/tamatebako/issues/1306)) ([6463a62](https://github.com/kufu/tamatebako/commit/6463a62384c1eb9a14b24e55638c37929e439736))
+
 ## [14.0.1](https://github.com/kufu/tamatebako/compare/eslint-config-smarthr-v14.0.0...eslint-config-smarthr-v14.0.1) (2026-04-19)